Implement basic writing support. Add test.
All parsed org files must be written out in identical format again. The
test verifies that. The Writer class will take an OrgElement and write
it into a QTextStream. OrgFile elements represent the top level elements
of an org file and will not create any content when written out. This
way OrgFile can be used to collect OrgElements that are to be written as
the top level elements of an org file.